usb: xhci: rename 'last_trb_seg' to 'end_seg' in struct 'xhci_td'
Rename 'last_trb_seg' to 'end_seg' so that it mirrors its counter part
'start_seg'.

A Transfer Descriptor (TD) in the xhci driver is a data structure that
represents a single transaction to be performed by the USB host controller.
This transaction is defined by TRBs from 'first_trb' in 'start_seg' to
'last_trb' in 'last_trb_seg' (which this patch changes to 'end_seg').

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>