usb: xhci: rework dma_in_range()
Function dma_in_range() searches for a DMA address within a segment ring,
starting from 'start_seg'. If the DMA address is found within a segment
and is positioned between 'start_trb' and 'end_trb', the function returns
the segment. If not, it returns 'NULL'. See ring example at the end.

Key enhancements include:

- Correctly, identify all wrap around cases. Previously, if the TRB was not
  in the Start/End segments, the dma_in_range() would return 'NULL'.
  e.g. [End - Start] -> [TRB]

- Instead of translating Start and End TRBs to DMA addresses, translate the
  DMA to TRB. This change reduces the amount of translations needed.

- Utilize 'end_seg' pointer as counterpart to 'start_seg', narrowing the
  search scope from start to end segment, improving efficiency.

- Prioritizing the most frequent scenario where the start and end
  segments are identical, and 'start_trb' precedes end_trb', by
  checking this case first for quicker resolution.

The start segment and start TRB checking has not been removed. Instead it
is moved to trb_in_td(). This should not ever be the case, both 'start_seg'
and 'start_trb' are set using 'ring->enq_seg' and 'ring->enqueue',
if either of them is 'NULL' something must have seriously gone wrong.
But I am not ready to pull the trigger and find out.

========================= Explanation =========================

There are the 15 unique possibilities:
 1.  [Start -> TRB -> End]
 2.  [End - Start -> TRB]
 3.  [Start -> End - TRB]
 4.  [TRB -> End - Start]
 5.  [End - TRB - Start]
 6.  [TRB - Start -> End]

 7.  [Start -> TRB]
 8.  [Start -> End]
 9.  [End - Start]
 10. [End - TRB]
 11. [TRB -> End]
 12. [TRB - Start]

 13. [Start] -> [End]
 14. [Start] -> [TRB] -> [End]
 15. [End] - [TRB] - [Start]

The complexity arises from the "wrap case," where the end loops back to
the start segment. To simplify, start and end segments are treated as
separate segments. The process is divided into three parts:
 1. Handle cases where TRB is in start segment and after the start TRB.
 2. Handle cases where TRB is in end segment.
 3. Check segments between, start and end segments.

In step 2 we only translate the DMA to TRB if start and end segments
are different.
Because step 1 handled all cases where TRB is after start TRB. In step 2,
the only valid option is if the TRB is before the end TRB.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>