xhci: Add helper to find trb from its dma address
Add a xhci_dma_to_trb() helper, and use it to find the transfer TRB
early in handle_tx_event() based on the dma address found in the
event TRB.

With this helper we can get rid of 'ep_seg' transfer TRB segment
variable which confusingly was used both as a boolean to indicate
if the transfer TRB is part of the next queued TD, and to actually
find the transfer TRB based on ep_seg and ep_trb_dma.

If the event points to the next queued td, i.e. trb matches
the next td then set a 'matched_td' pointer to that td instead of
using ep_seg as a boolean.

Proof of concept, not ready for upstream

NOT YET Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>