usb: xhci: adjust empty TD list is handling in handle_tx_event()
Introduce an initial check for an empty list prior to entering the while
loop. Which enables, the implementation of distinct warnings to
differentiate between scenarios where the list is initially empty and when
it has been emptied during processing (via skipping).

These adjustments not only refine the current implementation but also
facilitate future enhancements to the handle_tx_event() function.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>