usb: xhci: prevent skipping TDs past target TRB
An isoc ring, when set with the skip flag, attempts to bypass TDs until it
locates the TD containing its designated TRB. However, if the sought-after
TRB is absent from all TDs, all TDs are skipped. This behavior is
incorrect; only TDs preceding the target TRB should be skipped, as TDs
following the target TRB are intended for upcoming events.

To address this, introduce function trb_pos_idx(). This function aids in
determining a TRB's position relative to a TD, whether it is before,
within, or following the TD. It returns an index for the given TRB,
adjusted based on the dequeue TRB's position. Utilizing the target TRB's
index in comparison to a TD's start and end indices, the function
xhci_find_event_td() can now accurately assess whether a TD should be
skipped or not.

Note: in the next commit, all event types will be able to skip TDs. Thus,
      the function xhci_find_event_td() is not isoc ring specific.

============================= Example ====================================

Segment ring consisting of 3 segments (A,B) each containing 3 TRBs (0,1,2).

If dequeue TRB is A0:
Seg A: [0, 1, 2]
Seg B: [3, 4, 5]

If dequeue TRB is B1:
Seg A: [6, 7, 8]
Seg B: [9, 4, 5]

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>
Co-developed-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>