usb: xhci: remove argument 'xhci' from next_trb()
Argument 'xhci' (struct xhci_hcd *) is not used in next_trb().

Consequently the argument is also removed from functions:
td_to_noop() & sum_trb_lengths()

Argument 'ring' is also currently unused, but this is set to change
in a future patch. To prevent unnecessary back and forth changes the
argument is not removed.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>