usb: xhci: remove redundant and incorrect EINT bit write
Event Interrupt (EINT) is by default '0', xHC sets the bit to '1' when the
Interrupt Pending (IP) bit of any Interrupter transitions from '0' to '1'.
Software that uses EITN shall clear it prior to clearing any IP flags.
A race condition may occur if software clears the IP flags then clears the
EINT flag, and between the operations another IP '0' to '1' transition
occurs. In this case the new IP transition shall be lost.

The EINT flag does not generate an interrupt, it is simply a logical OR
of the IMAN register IP flag ‘0’ to ‘1’ transitions.
As such, it does not need to be cleared to clear an xHC interrupt.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>