usb: xhci: remove unnecessary re-allocations in xhci_reinit()
When resuming from hibernate (S4), the xhci driver currently frees all
memory and then re-allocates and re-initializes it. This process is
suboptimal, as much of the memory does not need to be freed and can
simply be re-initialized.

To address this, only free memory that is dynamically allocated during
runtime, such as virtual devices. Memory that does not change during
runtime, such as interrupters and the command ring, should only be
re-initialized.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>