usb: xhci: usb: xhci: rework xhci_free_segments_for_ring()
The segment list does not exist outside the context of the ring.
Therefore, it is clearer to pass the ring itself when freeing all segments
from a ring, rather than passing the ring list.

Additionally, rename the function to "xhci_ring_segments_free" to align
with the naming convention of xhci_segment_free().

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>