usb: xhci: improve xhci_set_interrupter_moderation()
The function sets the interrupt moderation interval through the IMODI
16-bit register, where the interval is represented in increments of 250
nanoseconds. For example, an IMODI register value of 16 is multiplied by
250 to get 4000 nanoseconds, resulting in an interrupt every ~1ms.

Currently, the function has the potential to fail, but this failure is not
handled, and the xhci driver proceeds as usual. Fix the issue by removing
the possibility of failure. Instead, the IMODI is set to the closest valid
value.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>