usb: xhci: set requested IMODI to the closest supported value
The function configures the Interrupt Moderation Interval (IMODI) via bits
15:0 in the Interrupt Moderation Register. The IMODI value is specified in
increments of 250 nanoseconds. For instance, an IMODI register value of 16
corresponds to 4000 nanoseconds, resulting in an interrupt every ~1ms.

Currently, the function fails when a requested IMODI value is too large.
Prevent this by setting IMOD to the closest supported value.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>