usb: xhci: improve xhci_set_interrupter_moderation()
The function sets the interrupt pace, through IMODI 16 bit register,
the interval represented in increments of 250. For example, IMODI
register value of 16 is multiplied by 250 to get 4000 nanoseconds,
which will result in an interrupt every ~1ms.

The function has the option of failing, currently nothing catches the
failure, instead xhci driver proceeds as usual. Fix this by removing
the option of failure, instead set the Interrupt moderation interval
to the closest valid value.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>