usb: xhci: move check for deferred or newly canceled TDs
If the Set TR Deq command fails, the hardware still has cached TDs that
should have been skipped. This means the xhci driver cannot return them,
and the TD list cannot be empty. The only scenario where the canceled TD
list could be empty is when the command completion event is successful,
allowing the skipped TDs to be returned. Therefore, move this check into
the 'COMP_SUCCESS' switch case.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>