usb: xhci: clear up next command retrieval
The 'list_first_entry()' function can be misleading in this context, as it
implies retrieving the first entry in a list, which may cause confusion.
On the other hand, 'list_next_entry()' clearly indicates that we are
retrieving the next entry relative to the current one.

Thus, replace the 'list_first_entry()' call with 'list_next_entry()'.
In this case, both calls achieve the same result, but the latter is clearer.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>