usb: xhci: clarify retrieval of next command
Using list_first_entry() in this context can be misleading, as it suggests
retrieving the first entry in the list rather than the entry following the
current one. This may cause confusion when reading the code.

Replacing it with list_next_entry() makes the intent explicit, it clearly
indicates that the next entry relative to the current one is being
accessed.

While both calls produce the same result in this case, list_next_entry()
improves code readability and maintainability.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>