usb: xhci: handle Set TR Deq Error due to Stopped/Error Endpoint state
A Set TR Deq Context State Error occurs when the Slot state is Disabled or
the Endpoint state is neither Stopped nor Error.

Add handling for cases where a Context State Error is received, but the
Slot state is Enabled and the Endpoint state is either:
 - Stopped: This situation arises from a race condition caused by ring
   restart latency after an old doorbell ring. The doorbell was rung
   before a Stop Endpoint Command was queued. However, the hardware
   completed the Stop Endpoint Command and began processing a Set TR Deq
   command queued by the xhci driver in the Stop Endpoint Command
   completion handler before the ring started running. Consequently,
   the xHC hardware processes the Set TR Deq command on a running ring,
   leading to a Context State Error. In this case, re-issue
   a new Set TR Deq command.

 - Error: The Endpoint state is correct, possibly due to a race condition
   similar to the Stopped state. In this case, re-issue
   a new Set TR Deq command.

All Transfer Descriptors (TDs) associated with the failed Set TR Deq
command are marked as 'TD_DIRTY'. This marking ensures that they will be
eligible for the subsequent Set TR Deq command.

xHCI specification, rev 1.2:
 Slot State		section 4.5.3
 Endpoint State		section 4.8.3
 Set TR Dequeue Pointer	section 4.6.10

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>