usb: xhci: handle Set TR Deq Error due to Disabled slot state
The Set TR Deq Slot Not Enabled error occurs when the Endpoint State is
disabled. In this state, the slot's Doorbell register is disabled, meaning
it cannot receive or handle Transfer Descriptors (TDs).

Since the Disabled and Enabled slot states share the same value and error
handling mechanism, fall through the "Slot Not Enabled" error switch case
into the "Context State Error" case. This ensures that the
"Slot Not Enabled" error is handled in the same manner as a "Context State
Error" due to a disabled slot.

xHCI specification, rev 1.2:
 Slot State		section 4.5.3

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>