usb: xhci: add xhci_dma_to_virt_trb()
This function is the counterpart to xhci_trb_virt_to_dma().
In certain cases, translating a DMA address to a TRB address is
simpler and more efficient.

For instance, in the upcoming rework of dma_in_range(), instead of
translating the start and end TRBs to DMA addresses, we can translate
the target DMA address to a TRB address. This approach reduces the
number of translations required.

Adding this function improves code familiarity and includes additional
error handling, making it a valuable addition.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>