xhci: dbc: Avoid event polling busyloop if pending rx transfers are inactive.
Event polling delay is set to 0 if there are any pending requests in
either rx or tx requests lists. checking for pending requests does
not work well for "IN" transfers as the tty driver queues requests
to the list with TRBs to the ring just to be prepared to receive
data from the host.

This causes unnecessary busylooping and cpu hogging.

Only set the event polling delay to 0 if there are pending tx "write"
transfers, or if it was less than 10ms since last active data transfer
in any direction.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>