usb: xhci: simplify Status register reset
Previously, the USBSTS register was cleared, and the Enable Interrupt
(EINT) bit was set to '1'. A mask (~0x1fff) was applied to prevent setting
reserved bits 31:13, although it inadvertently included reserved bits 7:5.
Fortunately, both ranges are marked as RsvdZ, meaning that setting these
bits to '0' has no effect.

Given this, the entire USBSTS register can be safely set to zero, except
for the EINT bit. This allows us to eliminate the USBSTS register read call
and the mask (~0x1fff), streamlining the code and removing a read() call.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>