usb: xhci: add comment to CMRT duplicate deletion
At the end of xhci_suspend() the Compliance Mode Recovery Timer is deleted.
Function xhci_resume() is called to restore the driver back from suspend,
in case the driver had lost power, e.g. hibernate, the CMRT timer is
deleted again. It is not obvious why, thus add a comment explaining the
reason.

In xhci_suspend(), the Compliance Mode Recovery Timer (CMRT) is deleted as
part of the suspend process. When xhci_resume() is invoked to restore the
driver from suspend, the CMRT timer is deleted again when power was lost,
such as during hibernation.

The reason behind this duplicate deletion is not obvious. Therefore, add
a comment to clarify the reasoning for removing the CMRT timer twice.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>