usb: xhci: simplify Status register reset
The current code clears the USBSTS register and then sets the Enable
Interrupt (EINT) bit to '1'. A mask (~0x1fff) is used to avoid modifying
reserved bits 31:13, but it also inadvertently includes reserved bits 7:5
and bit 1. These fields are defined as RsvdZ, so writing '0' to them has
no effect.

According to the xHCI specification, section 5.1.1:
"System software shall mask all reserved fields (Rsvd, RsvdP or RsvdZ) to
 '0' before evaluating a register or data structure value. This will enable
 current system software to run with future xHCI implementations that
 define the reserved fields."

Given this, it is safe to write zero to all bits of USBSTS except EINT.
This allows the mask and the register read to be removed, simplifying the
code and avoiding potential future issues.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>