usb: xhci: relocate Restore/Controller error check
The Restore Error or Host Controller Error occurs when the host controller
fails to resume after suspend. In such cases, the xhci driver is fully
re-initialized, similar to post-hibernation.

This error check is only relevant when the xhci driver attempts a normal
resume. Therefore, relocate the entire check into the existing normal
resume condition clause to ensure it is only executed when necessary.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>