usb: xhci: replace bit shifts with the BIT() macro in xhci-caps.h
Bit shift operations on a 32-bit signed can be problematic on some
architectures. Instead use BIT() macro, which returns a 64-bit unsigned
value. This ensures that the shift operation is performed on an unsigned
type, which is safer and more portable across different architectures.
Using unsigned integers for bit shifts avoids issues related to sign bits
and ensures consistent behavior.

Switch from 32-bit to 64-bit?
As far as I am aware, this does not cause any issues.
Performing bitwise operations between 32 and 64 bit values, the smaller
operand is promoted to match the size of the larger one, resulting in a
64-bit operation. This promotion extends the 32-bit value to 64 bits,
by zero-padding (for unsigned).

Will the change to 64-bit slow down the xhci driver?
On a 64-bit architecture - No. On a 32-bit architecture, yes? but IMO
the performance hit does not outweigh the readability and other benefits
of using BIT() macro.

Why not use test_bit()?
In my opinion, it's unnecessary for these macros. I would use it if the
macros were masks, then test_bit() fit in perfectly into the code directly.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>