usb: xhci: update Max Ports cap to align with the xHCI specification 1.2
Updates the maximum ports value from 127 to 255, aligning with the xHCI
specification version 1.2.

Macro HCS_MAX_PORTS() returns the maximum number of available ports by
reading bits 31:24 from the Host Controller Capability Register. However,
it incorrectly reads bits 30:24, limiting the maximum value to 127.
According to the xHCI specification, the correct maximum value is 255.

The original implementation of this macro, as introduced in commit [1],
which was implemented in accordance to xHCI specification 0.95. However,
the current specification, only documents changes as far back as version
0.96, which specifies the maximum as 255.

Additionally, commit [2] introduced the macro MAX_HC_PORTS, which also
defined the maximum ports as 127 but has never been utilized.

[1], commit 74c6874199af ("USB: xhci: Support xHCI host controllers and
			   USB 3.0 devices.")
[2], commit 0f2a79300a14 ("USB: xhci: Root hub support.")

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>