usb: xhci: improve Endpoint Context register debugging
Improve the debugging output for Endpoint Context registers in the xHCI
driver. The Endpoint Context registers consist of the following fields:
 bit 0 - Dequeue Cycle State.
 bits 3:1 - RsvdZ.
 bits 63:4 - TR Dequeue Pointer, is 16-byte aligned.

Instead of printing the entire 64-bit register as a single block, each
field is now printed separately using the appropriate printk() specifiers.
This approach enhances user interpretation and ensures the DMA address
format is displayed accurately.

In dbc_handle_xfer_event(), the Cycle information is intentionally
omitted as it is unnecessary for this context.

xHCI specification, section 6.2.3.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>