usb: xhci-dbgcap: convert Endpoint TRB DMA to CPU's native byte order
The host controller writes to the event ring dequeue and generates an
interrupt. When the dbgcap driver handles the event, it should convert the
host controller's written data to the CPU's native byte order. This ensures
proper handling of transfer events and prevents issues related to byte
order mismatches.

Endpoint TRB DMA conversion is already done in the xhci driver.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>