usb: xhci: allocate internal DCBAA mirror dynamically
Allocate the internal virtual device array from the heap instead of using
stack memory. The total allocated size remains the same, but this change
prepares the code for supporting a variable number of virtual device
pointers in the future.

The array is now allocated using kcalloc_node(), which zeroes the memory
automatically, making the explicit memset() call unnecessary.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>