usb: xhci: allocate DCBAA based on host controller max slots
Allocate the Device Context Base Address Array (DCBAA) according to the
maximum number of device slots supported by the host controller, instead
of always allocating the absolute maximum of 255 entries.

The valid slot range is 1 to 255, so (max_slots + 1) entries are allocated
to allow one-based slot indexing. This is also the reason why the macro
'MAX_HC_SLOTS' is defined as 256.

This change allows the xHCI driver to respect custom slot limits, reduces
unnecessary memory usage, and removes the obsolete "TODO" comment.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>