usb: xhci: change DCBAA allocation method
Instead of allocating the entire 'xhci_device_context_array' structure,
allocate only its component. This change moves the struct itself to the
stack while dynamically allocating the '__le64 *' array used for device
context pointers.

This refactor does not change the HC slots, but prepares the code to
support a variable number of HC slots in a future patch.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>