usb: xhci: change DCBAA allocation method
Instead of allocating the entire 'xhci_device_context_array' structure as
a whole, this patch embeds the structure within 'xhci_hcd' and allocates
only the '__le64 *' array for device context pointers separately.

This refactor does not alter the number of Host Controller (HC) slots but
simplifies memory management and lays the groundwork for supporting a
variable number of HC slots.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>