usb: xhci: limit number of interrupts to 128
The xHCI driver defines only 128 interrupter register slots, yet allows up
to 2047 interrupters. According to the xHCI specification, the maximum
valid number of interrupters and slots is 1024. These mismatches can lead
to out-of-range accesses and excessive memory use.

To address this, introduce a new variable, 'xhci->max_intrs', set to the
minimum of the register-reported value and 'MAX_HC_INTRS'. The interrupter
register slot array is defined with 1024 entries as a structural template,
as this will not increase memory usage, and runtime allocations are limited
to 'xhci->max_intrs'.

The Number of Interrupters (MaxIntrs) field occupies bits 18:8 of the
'HCSPARAMS1' register, which can yield a value up to 2047, although the
specification limits it to 1024. To solve this, the value is capped via
'MAX_HC_INTRS' preventing invalid access.

Macro 'MAX_HC_INTRS' is set to 128, although the xHCI specification allows
up to 1024 interrupters, raising this limit would provide no practical
benefit, as the driver currently only uses the primary interrupter (0),
and secondary interrupters (1+) are rarely, if ever, used in practice.
There has also not been any reports of usage beyond 128.

Summary:
 * Interrupter allocations are now limited to 128 from 2047.
 * Interrupter Register template slots are set to 1024 from 128.
 * Macro 'MAX_HC_INTRS' can be modified to set the interrupter limit.

Signed-off-by: Niklas Neronin <niklas.neronin@linux.intel.com>